a more serious attempt at consistency after slack convo; expect will mostly pass, but still have to go through tests
