+++ ext/EnzymeStaticArraysExt.jl
@@ -21,14 +21,10 @@ end
     end
 end
 
-@inline Enzyme._gradient_output(res, x::StaticArray) = similar_type(x)(res)
+Enzyme.gradient_output(df, x::StaticArray) = similar_type(x)(df)
 
-@inline _combine_size(::Size{s1}, ::Size{s2}) where {s1,s2} = Size{(s1..., s2...)}()
+_jacsize(::Size{s1}, ::Size{s2}) where {s1,s2} = Size(s1..., s2...)
 
-@inline Enzyme._jacobian_output(cols, col1::Number, x::StaticArray) = similar_type(x)(cols)
-
-@inline function Enzyme._jacobian_output(cols, cols1::StaticArray, x::StaticArray)
-    reshape(reduce(hcat, cols), _combine_size(Size(cols1), Size(x)))
-end
+Enzyme.jacsize(df1::StaticArray, x::StaticArray) = _jacsize(Size(df1), Size(x))
 
 end
